requirements.txt
streamlit>=1.35.0
reportlab>=4.2.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0
pdf2image>=1.17.0
pydantic>=2.7.0
typer>=0.12.0
streamlit-drawable-canvas>=0.9.3
